The fully setupped file folder and working flow model
