+++ Backend/corepack.cmd
@@ -1,0 +1,7 @@
+@SETLOCAL
+@IF EXIST "%~dp0\node.exe" (
+  "%~dp0\node.exe"  "%~dp0\node_modules\corepack\dist\corepack.js" %*
+) ELSE (
+  @SET PATHEXT=%PATHEXT:;.JS;=;%
+  node  "%~dp0\node_modules\corepack\dist\corepack.js" %*
+)
