+++ Flow.md
@@ -1,0 +1,8 @@
+There are different steps in making the backend.
+1:Choosing the tech stack:-
+ Backend framework- here we need the express.js
+ Database- We are using two things:-
+                                    primary Database:- postgre SQL is used for the structured databse but it cannot be heplful if the data changes(schema structure change),but since here no schema is changed so we have used the postgreSQL
+                                    Database tool: An object-relational mapper which helps us to interract with the the database like-:prisma
+ Frontend - React.js,HTML,CSS,JavaScript
+ Recommendation system- we are using the collaborative data of login data and data from different API's to recommend the user the best suitable course.
